fix: session close overflow (BIGINT token columns) + idle sweeper savepoint isolation + cache_read over-attribution (DWB-505, DWB-506)
- Widen dwb_sessions.total_tokens, hook_sessions.total_tokens, tracking_log.tokens
  to BIGINT (hand-written migration; autogenerate misses type changes). A 5.47B
  rollup overflowed INT and 500'd every close of session 65.
- Clamp the persisted rollup in close_session so a pathological value can never
  raise MySQL 1264 and abort a close.
- Isolate each idle-sweeper close in a SAVEPOINT: one failing session was rolling
  back the entire sweep batch every cycle, which is why the sweeper closed nothing
  after 7-17.
- Stop summing cache_read_input_tokens into total_tokens in transcript parsing:
  it re-counted the whole cached context every turn (97 percent of the inflated
  rollup). Totals are now input + output + cache_creation; cache_read remains in
  the breakdown. Historical rows were recomputed from stored breakdowns separately.
- Regression tests for overflow close, sweeper isolation, and attribution.

+++ backend/app/models/dwb_session.py
@@ -1,12 +1,12 @@
 # Path: app/models/dwb_session.py
 # File: dwb_session.py
 # Created: 2026-06-09
-# Purpose: DwbSession ORM model - passive user-bounded session for time + token rollup (DWB-335, DWB-346 headline, DWB-381 slash escape hatch, DWB-382 ai_classifier fallback [retired DWB-402, enum kept as tombstone], DWB-481 structured summary JSON column)
+# Purpose: DwbSession ORM model - passive user-bounded session for time + token rollup (DWB-335, DWB-346 headline, DWB-381 slash escape hatch, DWB-382 ai_classifier fallback [retired DWB-402, enum kept as tombstone], DWB-481 structured summary JSON column, DWB-505 total_tokens widened to BIGINT)
 # Caller: app/services/dwb_session.py (DWB-337), app/routers/dwb_sessions.py (DWB-338)
 # Callees: app/database.Base
 # Data In: DB rows
 # Data Out: DwbSession, DwbOpenMethod, DwbCloseMethod, DwbCloseReason
-# Last Modified: 2026-06-25
+# Last Modified: 2026-07-28
 
 import enum
 from datetime import datetime
@@ -111,7 +111,10 @@ class DwbSession(Base):
         Enum(DwbCloseReason), nullable=True
     )
 
-    total_tokens: Mapped[int] = mapped_column(Integer, nullable=False, default=0)
+    # BigInteger (DWB-505): rollup of linked hook_sessions.total_tokens. Was INT
+    # and overflowed on a 5.47B-token session (MySQL 1264), 500-ing every close;
+    # widened so a large rollup can never overflow the close UPDATE again.
+    total_tokens: Mapped[int] = mapped_column(BigInteger, nullable=False, default=0)
     total_time_seconds: Mapped[int] = mapped_column(
         Integer, nullable=False, default=0
     )
